requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.1
django-cors-headers==4.4.0
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
